# specs/001-vortex-backend/contracts/api.yaml
openapi: 3.0.3
info:
  title: Vortex GPU Audio Backend API
  description: High-performance audio processing backend with GPU acceleration
  version: 1.0.0
  contact:
    name: Vortex Audio Team
    email: support@vortexaudio.com
  license:
    name: Commercial License
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.vortexaudio.com/api
    description: Production server
paths:
...
  /audio/upload:
    post:
      summary: Upload audio file for processing
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Audio file to upload
      responses:
        '200':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AudioFileUploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
...
components:
  schemas:
    AudioFormat:
      type: object
      properties:
        extension:
          type: string
        mimeType:
          type: string
        name:
          type: string
        description:
          type: string
        maxSampleRate:
          type: integer
        maxBitDepth:
          type: integer
        maxChannels:
          type: integer
        isHighResolution:
          type: boolean
    AudioFileUploadResponse:
      type: object
      properties:
        fileId:
          type: string
        name:
          type: string
        format:
          $ref: '#/components/schemas/AudioFormat'
        size:
          type: integer
        status:
          type: string
          enum: [uploaded, processing, ready, error]
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string
...
    PayloadTooLarge:
      description: Request entity too large
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string
              maxSize:
                type: string
security:
  - ApiKeyAuth: []
...